+++ app/tasks/add.js
@@ -25,6 +25,12 @@ const optionsDefinitions = [
     alias: "h",
     type: Boolean,
     description: "Displays this usage guide."
+  },
+  {
+    name: "file",
+    alias: "f",
+    type: String,
+    description: "Path to vault file. This is useful if you want to manage project vault which other members of the team can import."
   }
 ];
 
@@ -55,8 +61,14 @@ class AddTask {
   async execute(cliArgs) {
     let options = this._taskUtils.parseCliArguments(cliArgs);
     this._taskUtils.testOption(options.user, "User is mandatory option.");
+    if (options.file) {
+      console.log(`Working with secure store on location ${options.file}`);
+    }else{
+      console.log(`Working with default secure store`);
+    }
 
-    let secureStoreCliCommon = await SecureStoreCliCommon.init();
+
+    let secureStoreCliCommon = await SecureStoreCliCommon.init(options.file);
     let secureStoreCnt = await secureStoreCliCommon.readSecureStore();
 
     let ac1 = await read({prompt: `Access code 1 for ${options.user} : `, silent: true});