+++ app/tasks/import.js
@@ -1,13 +1,15 @@
 const TaskUtils = require("../misc/task-utils");
 const SecureStoreCliCommon = require("../secure-store-cli-common");
+const {promisify} = require('util');
+const read = promisify(require("read"));
+const r2 = require("r2");
 
 const optionsDefinitions = [
   {
-    name: "user",
+    name: "source",
+    alias: "s",
     type: String,
-    defaultOption:true,
-    typeLabel: "{underline user uid}",
-    description: "Default option. UID of user(human or uuEE) or alias such as 'you'."
+    description: "Source file for import."
   },
   {
     name: "help",
@@ -25,12 +27,14 @@ const optionsDefinitions = [
 
 const help = [
   {
-    header: "rm command",
-    content: "Removes user credentials from the vault."
+    header: "import command",
+    content: "Import user credentials from source file."
   },
   {
     header: 'Synopsis',
-    content: '$ oidc-plus4u-vault rm {underline user}'
+    content: [
+      '$ oidc-plus4u-vault import -s {underline source file}'
+    ]
   },
   {
     header: 'Options',
@@ -38,7 +42,7 @@ const help = [
   }
 ];
 
-class RmTask {
+class ImportTask {
 
   constructor() {
     this._taskUtils = new TaskUtils(optionsDefinitions, help);
@@ -46,21 +50,25 @@ class RmTask {
 
   async execute(cliArgs) {
     let options = this._taskUtils.parseCliArguments(cliArgs);
-    this._taskUtils.testOption(options.user, "User is mandatory option.");
+    this._taskUtils.testOption(options.source, "Source is mandatory option.");
+    console.log(`Going to read source secure store ${options.source}`);
+    let sourceSecureStoreCliCommon = await SecureStoreCliCommon.init(options.source);
+    let sourceSecureStore = await sourceSecureStoreCliCommon.readSecureStore();
 
+    console.log(`Going to import credentials to vault.`);
     if (options.file) {
       console.log(`Working with secure store on location ${options.file}`);
     }else{
       console.log(`Working with default secure store`);
     }
 
-
     let secureStoreCliCommon = await SecureStoreCliCommon.init(options.file);
     let secureStoreCnt = await secureStoreCliCommon.readSecureStore();
-    delete secureStoreCnt[options.user];
-    secureStoreCliCommon.writeSecureStore(secureStoreCnt);
+
+    secureStoreCnt = {...secureStoreCnt, ...sourceSecureStore};
+    await secureStoreCliCommon.writeSecureStore(secureStoreCnt);
   }
 
 }
 
-module.exports = RmTask;
+module.exports = ImportTask;